chore: rename components
