+++ apps/meteor/client/components/PaginatedVirtualList/index.ts
@@ -1,0 +1,1 @@
+export { default as PaginatedVirtualList } from './PaginatedVirtualList';
