--- apps/meteor/client/components/VirtualList/index.ts
@@ -1,1 +1,0 @@
-export { default as PaginatedVirtualList } from './VirtualList';
